fix: move curate tests to its own file
